refactor(packages/contracts): reorganize auth and library contracts into request/response structure
Separate request and response schemas into dedicated modules for better organization
and maintainability. Update index exports to clearly distinguish between request and
response types with organized comments.

BREAKING CHANGE: Type names changed from `*Dto` suffix to `*Request`/`*Response` suffix.
Export paths changed from flat structure to `request/` and `response/` subdirectories.
