+++ packages/contracts/src/auth/index.ts
@@ -1,4 +1,11 @@
-export * from "./login";
-export * from "./logout";
-export * from "./refresh";
-export * from "./register";
+// Requests
+export * from "./request/login.request";
+export * from "./request/logout.request";
+export * from "./request/refresh.request";
+export * from "./request/register.request";
+
+// Responses
+export * from "./response/login.response";
+export * from "./response/logout.response";
+export * from "./response/refresh.response";
+export * from "./response/register.response";
