+++ packages/contracts/src/auth/request/logout.request.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const LogoutRequestSchema = z.object({}).strict();
+
+export type LogoutRequest = z.infer<typeof LogoutRequestSchema>;
