+++ packages/contracts/src/auth/request/refresh.request.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const RefreshRequestSchema = z.object({}).strict();
+
+export type RefreshRequest = z.infer<typeof RefreshRequestSchema>;
