+++ packages/contracts/src/library/index.ts
@@ -1,3 +1,9 @@
-export * from "./create-library";
-export * from "./get-library";
-export * from "./get-library-albums";
+// Requests
+export * from "./request/create-library.request";
+export * from "./request/get-library-albums.request";
+export * from "./request/get-library.request";
+
+// Responses
+export * from "./response/create-library.response";
+export * from "./response/get-library-albums.response";
+export * from "./response/get-library.response";
