+++ packages/contracts/src/library/request/create-library.request.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const CreateLibraryRequestSchema = z.object({}).strict();
+
+export type CreateLibraryRequest = z.infer<typeof CreateLibraryRequestSchema>;
