+++ packages/contracts/src/library/request/get-library.request.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const GetLibraryRequestSchema = z.object({}).strict();
+
+export type GetLibraryRequest = z.infer<typeof GetLibraryRequestSchema>;
