+++ packages/contracts/src/library/response/create-library.response.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+import { createApiResponseSchema } from "../../api";
+import { LibrarySchema } from "../../schemas";
+
+export const CreateLibraryResponseSchema =
+  createApiResponseSchema(LibrarySchema);
+
+export type CreateLibraryResponse = z.infer<typeof CreateLibraryResponseSchema>;
