+++ packages/contracts/src/library/response/get-library.response.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+import { createApiResponseSchema } from "../../api";
+import { LibrarySchema } from "../../schemas";
+
+export const GetLibraryResponseSchema = createApiResponseSchema(LibrarySchema);
+
+export type GetLibraryResponse = z.infer<typeof GetLibraryResponseSchema>;
